Move collapse border test cases to layout
